Add health handler
Extract health endpoint into a dedicated HealthHandler class and wire it up in Program.cs (app.MapGet("health", HealthHandler.HandleAsync)).
